samples/mcp/flight_booking/pyproject.toml
[project]
name = "flight-booking-demo"
version = "0.1.0"
description = "A2UI Flight Booking Demo with MCP and ADK"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp",
    "google-adk",
    "jsonschema",
    "a2ui",
    "starlette",
    "uvicorn",
    "sse-starlette",
    "json-rpc",
    "python-dotenv",
    "google-adk>=1.8.0",
    "google-genai>=1.27.0",
    "python-dotenv>=1.1.0",
    "litellm",
    "jsonschema>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.uv.sources]
a2ui = { path = "../../../a2a_agents/python/a2ui_extension", editable = true }

